+++ pulsr/fileCache.js
@@ -6,6 +6,7 @@
  * async-cache allows to significantly decrease I/O hits.
  */
 define(['async-cache', 'fs', 'conf'], function(AsyncCache, fs, conf) {
+    // async-cache to retrieve file stat information
     var stats = new AsyncCache({
         // options passed directly to the internal lru cache
         // indicating how many items to store in a fileCache
@@ -19,11 +20,30 @@ define(['async-cache', 'fs', 'conf'], function(AsyncCache, fs, conf) {
             // something that will be unique.
             // this method will only be called if it's not already in cache, and will
             // cache the result in the lru.
-            fs.stat(filePath, cb)
+            fs.stat(filePath, cb);
         }
-    })
+    }),
+    // async-cache to read the templates from layout files
+    // (main layout, page layouts, and pagelet layouts)
+    layouts = new AsyncCache({
+        // options passed directly to the internal lru cache
+        // indicating how many items to store in a fileCache
+        max: conf.app.cache.fileCache.layouts.max,
+        // refresh the cache every maxAge time
+        maxAge: conf.app.cache.fileCache.layouts.maxAge,
+        // method to load a thing if it's not in the cache.
+        // key must be unique in the context of this cache.
+        load: function (filePath, cb) {
+            // the key can be something like the path, or fd+path, or whatever.
+            // something that will be unique.
+            // this method will only be called if it's not already in cache, and will
+            // cache the result in the lru.
+            fs.readFile(filePath, 'utf8', cb);
+        }
+    });
 
     return {
-        stats: stats
+        stats: stats,
+        layouts: layouts
     };
 });