+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Pulsr Web App Framework Change Log
 
+## Version 0.1.0 December 18, 2012
+
+- Retrieve layouts using fileCache as well.
+- Added a test suit for a base controller.
+
 ## Version 0.0.9 December 18, 2012
 
 - A fileCache module which caches files stat information. Uses async-cache module.
@@ -8,8 +13,8 @@
 
 ## Version 0.0.8 December 18, 2012
 
-Use Streams to stream static files instead of reading into memory.
+- Use Streams to stream static files instead of reading into memory.
 
 ## Version 0.0.7 December 15, 2012
 
-Initial release
+- Initial release